Split game into sub components
